Extend requestFullResolutionOutput to Camera node
This commit adds a new flag to use the highest
resolutions in the requestFullResolutionOutput.

The main reason is backwards compatibility, as
the previous behavior was to not use 8k resolutions
since they weren't supported at the OS level on
RVC4.

Due to the very highest resolutions taking a lot
of resources, the boolean flag is now added to
only use the highest resolutions if explicilty
requested.